refactor: 스타일 파일 폴더구조 변경 (제가 작업한 것들만)
